refactor: :recycle: move `_get_nested_attr()` into internals
