+++ src/seedcase_sprout/internals/__init__.py
@@ -6,7 +6,7 @@ from .create import (
     _create_resource_properties_script_filename,
 )
 from .functionals import _map, _map2
-from .get import _get_iso_timestamp
+from .get import _get_iso_timestamp, _get_nested_attr
 from .read import _read_json
 from .to import _to_camel_case, _to_snake_case
 from .write import _write_json
@@ -18,6 +18,7 @@ __all__ = [
     "_create_resource_properties_script_filename",
     "_to_snake_case",
     "_get_iso_timestamp",
+    "_get_nested_attr",
     "_map",
     "_map2",
     "_read_json",
